changed type for entitys id

+++ src/main/java/com/example/demo/model/entity/Car.java
@@ -1,7 +1,6 @@
 package com.example.demo.model.entity;
 
 import lombok.Data;
-
 import javax.persistence.*;
 
 @Entity
@@ -10,12 +9,9 @@ import javax.persistence.*;
 public class Car {
 
     @Id
-//    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @GeneratedValue(strategy=GenerationType.SEQUENCE,
-            generator="car_id_seq")
-    @SequenceGenerator(name="car_id_seq",
-            sequenceName="car_id_seq", allocationSize=10)
-    private int id;
+    @GeneratedValue(strategy=GenerationType.SEQUENCE, generator="car_id_seq")
+    @SequenceGenerator(name="car_id_seq", sequenceName="car_id_seq", allocationSize=10)
+    private Long id;
     private String brand;
     private String model;
     private String vin;
